# protocol: optimize ConnectionID.String