This function is used by qlog, so it should be fast.

```
name                     old time/op    new time/op    delta
ConnectionIDStringer-16    65.2ns ± 4%    15.6ns ± 3%  -76.08%  (p=0.000 n=9+9)

name                     old alloc/op   new alloc/op   delta
ConnectionIDStringer-16     64.0B ± 0%     16.0B ± 0%  -75.00%  (p=0.000 n=10+10)

name                     old allocs/op  new allocs/op  delta
ConnectionIDStringer-16      3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.000 n=10+10)
```